GetPendingQuests refactored
